Moving towards config files to support different architectures
