+++ configs/dataset/unit_test.yaml
@@ -1,0 +1,10 @@
+dataset:
+  name: 'unit_test'
+  path: 'datasets/unit_test'
+  trbatch: 8
+  valbatch: 8
+  testbatch: 1
+  numworkers: 8
+  blur_kernel_size: [3, 7]
+  sigmas: [0.1, 1.5]
+  train_val_test_split: [0.7, 0.15, 0.15]
