+++ main.py
@@ -10,65 +10,28 @@ from src.model.model import Generator, Discriminator
 from src.train.train import PtTrainer
 from src.model.load_save import saveModelToDisk, loadModelFromDisk
 from src.utils.utils import Utils
-from src.utils.img_processing import Downsample
-from PIL import Image
-
-def show_images(trainer, loader, save_path):
-    
-    # Images, labels to device               
-    images, labels = next(iter(loader))
-    images_cuda = images.to('cuda')
-
-    # Forward pass
-    gens = trainer.generator(images_cuda)
-    gens = gens.detach().cpu()
-
-    # Normalize from [0, 1] to [0, 255]
-    images = (images * 255).byte()
-
-    # Normalize high resolution images from [-1, 1] to [0, 255]
-    gens = (((gens + 1) / 2) * 255).byte()
-    labels = (((labels + 1) / 2) * 255).byte()
-    
-    fig = plt.figure(figsize=(10,12), layout='compressed')
-    batch_size = len(images)
-    for i in range(batch_size):
-        plt.subplot(batch_size, 3, 3*i+1)
-        plt.imshow(images[i].permute(1, 2, 0))
-        if i == 0: plt.title("LR")
-
-        plt.subplot(batch_size, 3, 3*i+3)
-        plt.imshow(gens[i].permute(1, 2, 0))
-        if i == 0: plt.title("SR")
-
-        plt.subplot(batch_size, 3, 3*i+2)
-        plt.imshow(labels[i].permute(1, 2, 0))
-        if i == 0: plt.title("HR")
-
-    if save_path:
-        plt.savefig(os.path.join(save_path, "test_set_plot.png"))
-
-    plt.show()
+from torchsummary import summary
 
 # Trainer for models
 def train():
 
-    # Paths
+    # Get configs
     root_dir = os.getcwd()
+    dataset_name = "cifar"
+    model_config_name = "srgan-small-standard"
+    model_config, dataset_config = Utils.loadConfig(root_dir, model_config_name, dataset_name)
+
+    b1k_sz = model_config['model']['gen_block_1_kernel_size']
+    b1p_sz = model_config['model']['gen_block_1_padding_size']
+    n_resb = model_config['model']['gen_resid_blocks']
+    cc = model_config['model']['gen_conv_channels']
+    scale = model_config['model']['scale_factor']
+
+    g = Generator(b1k_sz, b1p_sz, n_resb, cc, scale)
+    summary(g, (3, 8, 8))
+
+    """
     model_dir = os.path.join(root_dir, 'models')
-
-    dataset_name = "CIFAR10"
-
-    # Load Hyperparameters
-    hyps = Utils.loadHypsFromDisk(os.path.join(os.path.join(model_dir, 'hyps'), dataset_name + '.txt'))
-
-    # Dataset
-    blur_kernel_size = (3,7)
-    sigma = (0.1,1.5)
-    batch_size_train = hyps["trbatch"]
-    batch_size_val = hyps["valbatch"]
-    num_workers = hyps["numworkers"]
-
     # Create dataset
     dataset_dir = os.path.join(os.path.join(root_dir, "datasets"), dataset_name.lower())
     if dataset_name == "ImageNet":
@@ -116,6 +79,7 @@ def train():
     plt.legend()
     plt.savefig(os.path.join(save_path, "loss_plot.png"))
     plt.show()
+    """
 
 # Sampling from GAN
 def eval():
@@ -162,81 +126,8 @@ def eval():
     # show example
     show_images(validator, test_loader, None)
 
-def showSamples():
-    """ Show samples from different parts of training
-    """
-    # Paths
-    root_dir = os.getcwd()
-    model_dir = os.path.join(root_dir, 'models')
-
-    dataset_name = "CIFAR10"
-
-    # Load Hyperparameters
-    hyps = Utils.loadHypsFromDisk(os.path.join(os.path.join(model_dir, 'hyps'), dataset_name + '.txt'))
-
-    # Dataset
-    blur_kernel_size = (3,7)
-    sigma = (0.1,1.5)
-    batch_size_train = hyps["trbatch"]
-    batch_size_val = hyps["valbatch"]
-    num_workers = hyps["numworkers"]
-
-    # Create dataset
-    dataset_dir = os.path.join(os.path.join(root_dir, "datasets"), dataset_name.lower())
-    if dataset_name == "ImageNet":
-        dataset = ImageNetDataset(dataset_dir, blur_kernel_size, sigma, batch_size_train, num_workers)
-    elif dataset_name == "CIFAR10":
-        dataset = CIFAR10Dataset(dataset_dir, blur_kernel_size, sigma, batch_size_train, num_workers)
-    
-    train_val_test_split = [.7,.15,.15]
-    train_dataset, val_dataset, test_dataset = random_split(dataset, train_val_test_split, generator=torch.Generator().manual_seed(42))
-    test_loader = DataLoader(test_dataset, batch_size=3, shuffle=False, num_workers=num_workers)
-
-    # Load Models
-    mse_model = f"SRGAN_epoch_{5}_scale_{hyps['scale']}_k3_21"
-    mse_path = os.path.join(model_dir, "cifar/generator_block12_k3/" + mse_model)
-    gmse, _ = loadModelFromDisk(mse_path, hyps)
-
-    feat_model = f"SRGAN_epoch_{5}_scale_{hyps['scale']}_k3_21_feat"
-    feat_path = os.path.join(model_dir, "cifar/generator_block12_k3/" + feat_model)
-    gfeat, _ = loadModelFromDisk(feat_path, hyps)
-
-    distf_model = f"SRGAN_epoch_{5}_scale_{hyps['scale']}_k3_7_feat"
-    distf_path = os.path.join(model_dir, "cifar/generator_block12_k3/" + distf_model)
-    gdistf, _ = loadModelFromDisk(distf_path, hyps)
-
-    lr, hr = next(iter(test_loader))
-    mse_gens = gmse(lr).detach()
-    feat_gens = gfeat(lr).detach()
-    distf_gens = gdistf(lr).detach()
-
-    #lr = (lr * 255).byte()
-    # Concatenate images together
-    images = [lr, hr, mse_gens, feat_gens, distf_gens]
-
-    # Plotting the concatenated images
-    fig, axes = plt.subplots(3, 5, figsize=(8, 5))
-    plt.tight_layout(pad=0.5, w_pad=0.5, h_pad=0.5)
-    axes[0,0].set_title("LR")
-    axes[0,1].set_title("HR")
-    axes[0,2].set_title("MSE-Only")
-    axes[0,3].set_title("Feat")
-    axes[0,4].set_title("Dist Failure")
-    fig.patch.set_facecolor('white')
-
-    for i in range(lr.shape[0]):
-        for j in range(len(images)):
-            img = images[j][i].permute((1, 2, 0))  # Convert (C, H, W) to (H, W, C)
-            # Normalize high resolution images from [-1, 1] to [0, 255]
-            if j != 0:
-                img = (((img + 1) / 2) * 255).byte()
-            axes[i,j].imshow(img)
-            axes[i,j].axis('on')
-
-    plt.tight_layout()
-    plt.show()
 
 if __name__ == "__main__":
-    #train()
+    train()
     #eval()
-    showSamples()
+    #showSamples()
