--- models/hyps/ImageNet.txt
@@ -1,7 +1,0 @@
-epochs=5$i
-scale=4$i
-lr=2e-5$f
-dropout=0.3$f
-trbatch=8$i
-valbatch=8$i
-numworkers=8$i
