+++ src/model/model.py
@@ -5,7 +5,7 @@ import torchvision.transforms.functional as F
 
 class ResidualBlock(nn.Module):
     '''
-    This is Residuel Block in the Generator Network
+    This is Residuel Block in the generator network
     Note that spatial depth is maintained with padding, 
     this is essential for the residual layers
     '''
@@ -27,31 +27,14 @@ class ResidualBlock(nn.Module):
         
         return output
 
-class DisBlock(nn.Module):
-    '''
-    This is the main structure in Discriminator network
-    '''
-    def __init__(self, channel, scale=1, stride=1):
-        super(DisBlock, self).__init__()
-        self.conv = nn.Conv2d(channel, channel * scale, kernel_size=3, padding=1, stride=stride)
-        self.bn = nn.BatchNorm2d(channel * scale)
-        self.lekrelu = nn.LeakyReLU(0.2)
-    
-    def forward(self, x):
-        x = self.conv(x)
-        x = self.bn(x)
-        x = self.lekrelu(x)
-
-        return x
-
-class SubPixel(nn.Module):
+class UpSample(nn.Module):
     '''
     Upsample the image spatially
     For scale factor = r, conv (c_in, c_out r**2)
     Reshape output of conv to shape: (c_in, h * r, w * r)
     '''
     def __init__(self, channel_in, scale):
-        super(SubPixel, self).__init__()
+        super(UpSample, self).__init__()
         self.up_scale = int(math.sqrt(scale))
         self.conv = nn.Conv2d(channel_in, channel_in * self.up_scale**2, kernel_size=3, padding=1)
         self.pixshuff = nn.PixelShuffle(self.up_scale)
@@ -67,50 +50,66 @@ class Generator(nn.Module):
     '''
     SRGAN Generator
     '''
-    def __init__(self, scale):
+    def __init__(self, block_1_k_size, block_1_padding, num_resid_blocks, conv_channels, \
+                 scale):
         super(Generator, self).__init__()
 
-        self.block1 = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=9, padding=4),
+        # Initial feature extraction
+        # Kernel size should be tuned to size of images
+        self.fextraction = nn.Sequential(
+            nn.Conv2d(3, 64, kernel_size=block_1_k_size, padding=block_1_padding),
             nn.PReLU()
         )
-        self.block2 = ResidualBlock(64)
-        self.block3 = ResidualBlock(64)
-        self.block4 = ResidualBlock(64) 
-        self.block5 = ResidualBlock(64)
-        self.block6 = ResidualBlock(64)
-        self.block7 = ResidualBlock(64)
-        self.block8 = ResidualBlock(64)
-        self.block9 = nn.Sequential(
-            nn.Conv2d(64, 64, kernel_size=3, padding=1),
-            nn.BatchNorm2d(64)
-        )
-        # do element-wise sum before passing into block 10 (in forward)
 
-        self.block10 = SubPixel(64, scale)
-        self.block11 = SubPixel(64, scale)
-        self.block12 = nn.Sequential(nn.Conv2d(64, 3, kernel_size=3, padding=1))
+        # Stack residual block
+        # The number of residual blocks is the main difference across network sizes
+        self.resblocks = nn.Sequential(*[ResidualBlock(conv_channels) for i in range(num_resid_blocks)])
+
+        # Consolidate features from residual extraction
+        self.fconsolidation = nn.Sequential(
+            nn.Conv2d(conv_channels, conv_channels, kernel_size=3, padding=1),
+            nn.BatchNorm2d(conv_channels)
+        )
+
+        # Upsample to HR image size
+        self.up1 = UpSample(conv_channels, scale)
+        self.up2 = UpSample(conv_channels, scale)
+
+        # Final conv to clean up upsampled feature representation
+        self.convout = nn.Sequential(nn.Conv2d(conv_channels, 3, kernel_size=3, padding=1))
 
     def forward(self, x):
-        x = self.block1(x)
-        output = self.block2(x)
-        output = self.block3(output)
-        output = self.block4(output) 
-        output = self.block5(output)
-        output = self.block6(output)
-        output = self.block7(output)
-        output = self.block8(output)
-        output = self.block9(output)
+        x = self.fextraction(x)
+        out = self.resblocks(x)
+        out = self.fconsolidation(out)
+        
         # Element wise concat
-        output = output + x
-        output = self.block10(output)
-        output = self.block11(output)
-        output = self.block12(output)
+        out = out + x
 
-        # Output logits
-        output = torch.tanh(output)
+        out = self.up1(out)
+        out = self.up2(out)
+        out = self.convout(out)
+        
+        # Align with [-1, 1] input scale
+        out = torch.tanh(out)
+        return out
 
-        return output
+class DisBlock(nn.Module):
+    '''
+    Block for discriminator
+    '''
+    def __init__(self, channel, scale=1, stride=1):
+        super(DisBlock, self).__init__()
+        self.conv = nn.Conv2d(channel, channel * scale, kernel_size=3, padding=1, stride=stride)
+        self.bn = nn.BatchNorm2d(channel * scale)
+        self.lekrelu = nn.LeakyReLU(0.2)
+    
+    def forward(self, x):
+        x = self.conv(x)
+        x = self.bn(x)
+        x = self.lekrelu(x)
+
+        return x
 
 class Discriminator(nn.Module):
     '''
